Simplify terminal interfaces, remove changelog
In order to simplify the interface of the terminal I've moved all the
formatting, color, and "ui" functionality out into traits and builder
objects. This gets rid of the annoying "format_x" functions and all the variations
of the draw_box functions, and allows for adding new formatting options
without modifying the terminal interface.

Also to save me the work I've removed the changelog. Detailed commits should
serve just as well.